// src/contexts/ads/ads.types.ts
export enum AdAccountMemberRole {
  OWNER = 'owner',
  ADMIN = 'admin',
  ANALYST = 'analyst',
}

export enum AdCampaignStatus {
  DRAFT = 'draft',
  SCHEDULED = 'scheduled',
  ACTIVE = 'active',
  ENDED = 'ended',
  PAUSED = 'paused',
}

export enum AdCampaignPlacement {
  BANNER = 'banner',
  SIDEBAR = 'sidebar',
}